+++ example/lib/example1/example1.dart
@@ -1,0 +1,10 @@
+import 'package:demoflu/demoflu.dart';
+import 'package:demoflu_example/example1/example1_widget.dart';
+
+class Example1 extends Example {
+  Example1()
+      : super(
+            widget: MainWidget(),
+            codeFile: 'lib/example1/example1_widget.dart',
+            resizable: true);
+}