+++ example/lib/example1/example1_widget.dart
@@ -1,12 +1,4 @@
-import 'package:demoflu/demoflu.dart';
 import 'package:flutter/material.dart';
-
-class Example1 extends Example {
-  @override
-  Widget buildMainWidget(BuildContext context) {
-    return MainWidget();
-  }
-}
 
 class MainWidget extends StatelessWidget {
   @override