+++ example/lib/example2/example2.dart
@@ -1,0 +1,10 @@
+import 'package:demoflu/demoflu.dart';
+import 'package:demoflu_example/example2/example2_widget.dart';
+
+class Example2 extends Example {
+  Example2()
+      : super(
+            widget: MainWidget(),
+            codeFile: 'lib/example2/example2_widget.dart',
+            consoleEnabled: true);
+}