+++ example/lib/example2/example2_widget.dart
@@ -1,13 +1,6 @@
 import 'package:demoflu/demoflu.dart';
 import 'package:flutter/material.dart';
 
-class Example2 extends Example {
-  @override
-  Widget buildMainWidget(BuildContext context) {
-    return MainWidget();
-  }
-}
-
 class MainWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {