+++ lib/demoflu.dart
@@ -3,4 +3,3 @@ library demoflu;
 export 'src/demoflu_app.dart';
 export 'src/demo_menu_item.dart';
 export 'src/example.dart';
-export 'src/example_builder.dart';