+++ lib/src/demo_menu_item.dart
@@ -1,25 +1,16 @@
-import 'package:demoflu/src/example_builder.dart';
-import 'package:flutter/material.dart';
+import 'package:demoflu/src/example.dart';
 
 /// Represents a menu item in the app menu.
 class DemoMenuItem {
   /// Builds a [DemoMenuItem].
   DemoMenuItem(
       {required this.name,
-      this.builder,
-      this.resizable,
-      this.codeFile,
-      this.maxSize,
-      this.consoleEnabled,
+      this.example,
       this.italic = false,
       this.indentation = 1});
 
   final String name;
-  final ExampleBuilder? builder;
-  final String? codeFile;
-  final bool? resizable;
-  final Size? maxSize;
-  final bool? consoleEnabled;
   final bool italic;
   final int indentation;
+  final AbstractExample? example;
 }