+++ lib/src/demoflu_tab_widget.dart
@@ -1,0 +1,18 @@
+import 'package:demoflu/src/example.dart';
+import 'package:flutter/cupertino.dart';
+
+
+class DemofluTabWidget extends StatelessWidget {
+
+  final Example example;
+
+  const DemofluTabWidget({Key? key, required this.example}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    // TODO: implement build
+    throw UnimplementedError();
+  }
+
+}
+