+++ lib/src/example_builder.dart
@@ -1,3 +1,0 @@
-import 'package:demoflu/src/example.dart';
-
-typedef ExampleBuilder = Example Function();