+++ lib/src/internal/switch_view/switch_view_toggle_buttons.dart
@@ -23,35 +23,21 @@ class SwitchViewToggleButtons extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-        child: Center(
-            child: ToggleButtons(
-          direction: Axis.horizontal,
-          onPressed: (int index) => onChange(SwitchViewType.values[index]),
-          borderRadius: const BorderRadius.all(Radius.circular(8)),
-          borderColor: Colors.blueGrey[200],
-          selectedBorderColor: Colors.blueGrey[200],
-          selectedColor: Colors.white,
-          fillColor: Colors.blueGrey[600],
-          color: Colors.blueGrey[400],
-          constraints: const BoxConstraints(
-            minHeight: 25,
-            minWidth: 80,
-          ),
-          isSelected: _selectedViews,
-          children: views,
-        )),
-        padding: EdgeInsets.all(16),
-        decoration: BoxDecoration(
-            boxShadow: [
-              BoxShadow(
-                color: Colors.black.withOpacity(0.5),
-                spreadRadius: 3,
-                blurRadius: 3,
-                offset: Offset(0, 0), // changes position of shadow
-              ),
-            ],
-            color: Colors.blueGrey[50],
-            border: Border(bottom: BorderSide(color: Colors.blueGrey[700]!))));
+    return ToggleButtons(
+      direction: Axis.horizontal,
+      onPressed: (int index) => onChange(SwitchViewType.values[index]),
+      borderRadius: const BorderRadius.all(Radius.circular(8)),
+      borderColor: Colors.blueGrey[200],
+      selectedBorderColor: Colors.blueGrey[200],
+      selectedColor: Colors.white,
+      fillColor: Colors.blueGrey[600],
+      color: Colors.blueGrey[400],
+      constraints: const BoxConstraints(
+        minHeight: 25,
+        minWidth: 80,
+      ),
+      isSelected: _selectedViews,
+      children: views,
+    );
   }
 }